requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
blis==0.7.11
braceexpand==0.1.7
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
datasets==3.0.1
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
dill==0.3.8
einops==0.8.0
en-core-web-lg @ file:///home/user/MLLMs/data/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
entrypoints==0.4
executing==2.1.0
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
# Editable Git install with no remote (minGPT==0.0.1)
-e /home/user/MLLMs/minGPT-master
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.9.11
requests==2.32.3
rich==14.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==7.1.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.2
sympy==1.13.3
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.9
tokenizer==3.4.5
tokenizers==0.13.3
torch @ file:///home/user/MLLMs/data/torch/torch-2.4.1%2Bcu118-cp39-cp39-linux_x86_64.whl#sha256=82309da0cce45cf61eb48b0567c7d080992d8ba98264da128ee1d858fac5dd75
torchvision @ file:///home/user/MLLMs/data/torch/torchvision-0.19.1%2Bcu118-cp39-cp39-linux_x86_64.whl#sha256=029df98e1d7b7b2064d2c1014e3e1594fc00f6777965b305934c774011aaaea8
tornado==6.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.0.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
utils==1.0.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
wrapt==1.17.2
xxhash==3.5.0
yarl==1.13.0
zipp==3.20.2
